Rename ccode.py, fcode.py, and cxxcode.py to c.py, fortran.py, and cxxcode.py
This is unfortunately a backwards incompatible change for anyone who imports
the printer classes from these submodules. However, it the only way to avoid
importing sympy.codegen at the top-level. The reason is this:

- If the module names share the same name as the function names, like ccode.py
  and ccode(), then whichever is imported last will be set as
  sympy.printing.ccode.

- sympy/printing/ccode.py (now sympy/printing/c.py) imports several things
  from sympy.codegen at the top-level. These are used in the class definition,
  so they are not easy to remove. Thus, if it is imported at the top-level,
  then so will sympy.codegen. We want to avoid this, as sympy.codegen should
  be a completely optional submodule and importing it at the top-level import
  time just adds to the import time.

- As a result, we have moved, the functions ccode(), print_ccode(), and so on
  to sympy.printing.codeprinter, so that they can be imported without
  importing sympy.printing.ccode.  However, this means that at first import,
  sympy.printing.ccode will be the function. But then as soon as the module
  sympy.printing.ccode is imported, it will replace this with the module,
  meaning that 'from sympy.printing import ccode' will start returning the
  module instead of the function.

There are only two ways to avoid this behavior. One is to make sure that
sympy.printing.ccode is imported before the ccode() function, as whichever one
is imported last will be sympy.printing.ccode. However, this requires
importing sympy.codegen at the top-level. The other way is to rename the
module, so that it doesn't conflict with the function name.

Making module names not conflict with function names is a good idea in
general, as it avoids this problem. However, it does mean a backwards
comparability break for any existing module. I don't know how we should
approach this for other such modules in SymPy. Currently, only these are
changed, as it is required to avoid the top-level sympy.codegen import.