core: introduce add and mul functions
Introduce add and mul functions as extensible alternatives to calling
Add or Mul directly. It is intended that these functions be used in core
so that user defined types can customise the behaviour of core routines.
This commit only introduces the functions but future commits will
make use of them in core and it is hoped that this can be used to
eliminate the experimental post-processors mechanism that is
currently used for matrix expressions. This new mechanism is also still
experimental and might be changed or removed later.

When different types of arguments are applied, the handler method of
type which presides the operation is used. For this, _op_priority of
passed types are checked first. If multiple types having same
_op_priority are returned, the function determines the priority relation
using multiple dispatch.

For more detailed information, refer to:
sympy.core.operations.AssocOpDispatcher.