core: make _sympify reject Basic subclasses
Previously a Basic subclass could pass through _sympify e.g.:

  >>> _sympify(sin) is sin
  True

This was problemantic because _sympify is used to validate and prepare
args for many Basic subclasses which should only have Basic args. With
this change using _sympify with a Basic subclass will now raise an
error:

  >>> _sympify(sin)
  ...
  SympifyError: SympifyError: sin

This should mean that the output of _sympify is guaranteed to be of type
Basic.