Replace algorithm of primitive_element for ex=True.
The algorithm currently used by primitive_element for ex=True is
based on the groebner function. That does not work reliably if
the input extension elements are not linearly disjoint (see e.g.
https://github.com/sympy/sympy/issues/14117). This commit implements
the same iterative method that is used for the ex=False case extended
by including representations of field elements.