perf(core): make sympify faster for Basic subclasses.
The behaviour of sympify was changed in

  https://github.com/sympy/sympy/pull/20128

so that sympifying a Basic subclass would be an error when strict=True.
That change made the codepath for calling e.g. _sympify(Basic) slower as
more checks would be done before returning. This commit adds an early
raise for the case of calling _sympify(a) where a is a Basic subclass.