Improve apply_support method
Currently if you add a support at a location defined by decimal value, such as

    b.apply_support(0.2,'fixed') # fixed, pin, roller

and you try to catch the symbols created by the method with

    M,R = symbols('M_0.2, R_0.2')

the code fails because `apply_support`, when converting numerical value of location to a string in order to create symbols, maintains the trailing zeros. In order to make the code work the user should write something like

    M,R = symbols('M_0.200000000, R_0.200000000')

which is clearly counter-intuitive (how many zeros?).

With this simple fix trailing zeros are stripped from created symbols so that the code that the user is expecting to work runs correctly.