Convert_to(): Added a dimension check option that verifies unit consistency (Issue #18368):
If inconsistent units are used as target_units in the convert_to() function, output is  sometimes non-meaningful (Issue #18368).  A consistency check was added that verifies the output expression dimensions are consistent with the input expression dimensions.  This consistency check can be turned on and off with an optional parameter.  The consistency check currently does not include units in exponents, so sometimes the issue of a meaningless answer still persists for the exponent case.  The check could easily be expanded to include exponents by verifying the units in an exponent resolve to a dimensionless parameter; however, such a check should perhaps be integrated into the Pow() function operations itself to disallow exponent operations with units in general.

Example:
In [1] from sympy.physics.units import *

#Without dimension check
In [2] convert_to(watt, joule, checkUnits = False)
Out [2]  10**(1/3)*joule**(11/9)/10

#With dimension check
In [3] convert_to(watt, joule)
Out [3] Exception: Inconsistent unit conversion.  Verify target_units are compatible with input expr.

#Dimension check with units in exponent
In [4] convert(2**watt, joule)
Out [4] 2**(10**(1/3)*joule**(11/9)/10)