sets: addressed issue 20364 and made Rationals.contains(float) indeterminate
Currently, sympy.Rationals.contains(0.5) evaluates to False.
This patch instead evaluates the expression to indeterminate
based on the following logic: a float represents an approximation
of an underlying number that could be either rational or irrational.
In relation with this change, I have also adjusted a test case
that currently asserts sympy.Rationals.contains(1.0) evaluates
to False. This test was changed to assert that the expression
above evaluates to indeterminate instead of False or True.