feat(polys): make FiniteExtension a proper domain
Allow division of an ExtensionElement over a ring but only where the
divisor is an invertible constant (unit) of the ground domain.

For example in QQ[y][x]/(x**2 - 2) we can divide by any nonzero element
of QQ (the units of QQ[y]). In ZZ[x]/(x**2 - 2) we can only divide by 1
or -1.

If the ground domain for a FiniteExtension is a field then it is assumed
that the extension is also a field although this might not hold if the
quotient polynomial is not irreducible (this is not checked). If the
ground domain is a field then division by any nonzero element is
allowed.

This applies in particular to nested extensions e.g.:
  K = QQ[x]/(x**2 - 2)[y]/(y**2 - x)
Here the ground domain is QQ[x]/(x**2 - 2) which is assumed to be a
field because QQ is. It is therefore assumed that K is a field as well.
Division by any nonzero element of K is then permitted and considered
exact so that f % g is always zero.