fix(polys): fix FiniteExtension.exquo for gmpy
The FiniteEtension.exquo method was using self.domain (e.g. QQ) rather
than self.ring (QQ[x]) in order to compute division with the rep. This
somehow worked when gmpy was not installed but fails as should be
expected when gmpy is installed. This commit ensures the correct domain
is used.