maint(ci): remove coverage test jobs from Travis
This commit removes the coverage testing jobs from the Travis config. At
the time of this commit sympy is migrating from Travis to Actions and
Travis is still running just to see if it picks up on anything that
isn't picked up by Actions. The coverage jobs are failing on Travis due
to timeouts that are actually caused by Travis itself slowing down. This
commit removes the coverage tests so that the build will pass.