Fix pretty printing of midbar for ImageSet, ConditionSet, and ComplexRegion
Previously, it just printed | as a single line. Now the | is as tall as the
parentheses, and it uses the connecting Unicode character instead of | for the
Unicode pretty printing.

There is still a TODO here, which is to fix the baseline calculation for these
things so that the baselines are separate to the left and right of the |. I'm
not sure how to do that, though, so I've left it open.

I also made it so that the ASCII printer for these sets prints the large {
instead of { in a single line, as it looks odd with the large | otherwise (IMO
this shouldn't even be an option, but it's there).