Don't try to parse methods as implicit multiplications in parse_expr
For example x.y(z) should not be parsed as x.y*z.

Attributes of an object are almost never expressions that would be multiplied.
Methods are far more likely. Unlike for top-level names, we can't easily guess
what methods an object will have since we don't know its type, so we just
disable all implicit multiplications if the first name is a method.

Fixes #21020.