feat(polys): allow DomainMatrix to be dense or sparse
The DomainMatrix class was originally introduces with DDM as its
internal representation representing a dense internal format. The SDM
sparse format was added later and made to be the internal format of DDM
always. This commit makes it possible DomainMatrix to use either SDM or
DDM as its internal format in the constructor by passing fmt="dense" or
fmt="sparse".