Fixed bug with surface_color
Even though the documentation clearly states that `surface_color` should be a function,
if a user erroneously typed a float number, an error would have been raised, because
the `nb_of_points` attribute is not defined for Surface series classes.
With this change, the error has been solved by using the two attributes defined in
those classes.