fix(core): fix pickling of Basic to use __getnewargs__
This ensures that pickling of Basic subclasses is based on
__getnewargs__ with the exception of Symbol that uses __getnewargs_ex__
for keyword argument support. The changes here fix pickling using
protocol 2 or higher but remove support for pickling using protocol 0
or 1. Protocol 2 was introduced in Python 2.3.

A number of classes needed to rearranged or tidied up to accommodate for
this including Singleton and some of the Singleton class such as Reals.