Plotting - Improved custom labelling
As mentioned in issue 21300, the procedure to set custom labels to the expressions being plotted
was tedious. This commit add a new optional parameter to the plot interface, allowing to quickly
set custom labels to the expressions being plotted. For example, we can now write:

```
var("x, y")
plot((cos(x), "$f_{1}$"), (sin(x), (x, -2, 2), "f2"), (x / 10, (x, -6, 6)), legend=True)
```

Here, the first expression will be plotted over the default range `(x, -10, 10)` and will have a
label containing Latex code. The second expression will be plotted over the custom range
`(x, -2, 2)` having a string label. The last expression will be plotted over the custom range
`(x, -6, 6)` having the default label set to its string representation.

Back-compatibility has been an important thing to maintain. However, a few adjustments had to be
done. For example, previously it was possible to set the label of a single expression with the
`label` keyword:

```
plot(sin(x), label="test", legend=True)
```

With this commit, that keyword has been phased out (it doesn't do anything anymore, look at the
`_remove_label_key` function), and the docstrings has been updated with all the necessary
informations to correctly set custom labels.

From an implementation point of view, the old `check_arguments` function has been deprecated:
it was easier to come up with my own implementation in order to accomodate the new custom label
option, which can be found in the `_check_arguments` function. The latter has the same signature
of the former, and returns objects in the same format (with the added label string). Quite a few
tests have been implemented to make sure the new function produces the expected results.

All plot functions are going to sympify the provided arguments. However, when attempting to
sympify a string, it would either return an object of type Symbol, or it would raise an error
if any strange characters are present (/, {, }, $, ...). In order to solve this problem I had to
implement the `_plot_sympify` function, which is going to recursively sympify the provided
arguments, excluding the strings.