make symbolic Range canonical
Range tries to be as lenient as possible in accepting symbolic limits
but must then be careful to make no assumptions when being asked
for elements of that Range.

e.g. Range(x,x+5)[0] is not x unless we know that x is an integer;
we also do not know that the size is 5 unless we know if x is an
integer. If x == oo then this would be a null range.