Added first point to the parametric adaptive algorithm
In the previous commit I modified the code to return lists of points instead of list of segments,
but I forgot to insert the first point to the resulting lists. Hence, when the parametric_plot
would result in a missing point at the start of the curve. This commit fixes it.