solveset._invert_real: Fix irrational exponent case
Original code would raise ValueError on 'not base.is_positive'
assuming that the only situation for this case is when the
exponent is irrational. Issue 21236 showed that this case is
reached even when exponent is a Float. Also, the 'not base.is_positive'
is invalid for three-valued logic. solveset removes assumptions before
calling invert causing 'not base.is_positive' to be True regardless.