refactor(polys): move gmpy handling code to sympy.external
This commit moves the PythonRational class to a new sympy.external.gmpy
module. The new class PythonMPQ is defined to be compatible with gmpy2's
mpq type. This commit also drops support for integrating with gmpy
(1.x) since gmpy2 was released 10 years ago at this time.

It is now possible to do
   from sympy.external.gmpy import MPQ
which will give either PythonMPQ or gmpy's mpq type. A number of other
functions from gmpy2 have been moved into the sympy.external.gmpy
compatibility module so that they can be reused in other places. This
should make it possible to eliminate many of the polys domains that have
multiple versions for when gmpy is or is not installed.