refactor(polys): combine ZZ_gmpy and ZZ_python into ZZ
This commit makes it so that the same domain object ZZ is used
regardless of whether or not gmpy2 is installed. The ZZ_gmpy and
ZZ_python domains are merged into the ZZ superclass domain. The same is
done for QQ.

A subsequent commit could now delete the python* and gmpy* domain files
and all of the from_ZZ_python etc domain methods.

Some documentation formatting in polys is fixed as well.