Added size property to Union class
I added implementation for the size of the Union of two disjoint
Ranges. Originally, size is a property of Ranges. For example,
Range(0,5,1).size outputs 5. However, if we take the Union of
two disjoint Ranges: Union(Range(0,5,1),Range(5,8,1)), SymPy
did not know how to calculate the size of this set. This is why
I added the size property to the Union class. As an example,
Union(Range(0,5,1),Range(5,8,1)).size now outputs 8, as
expected.