Code Refactoring to improve readability
Previously, `get_points` methods was called when `adaptive=False`, hence it was used to
uniformly sample the line series (it returned lists of coordinates).

Now, `get_points` is always called. It will use a uniform or adaptive sampling depending
on the value of the `adaptive` attribute. As before, it will return list of coordinates,
hence back-compatibility is maintained.