Simplified the keyword arguments to set number of discretization points
As mentioned in issue #21282 , previously a user has to remember to use `nb_of_points`,
`nb_of_points_x`, `nb_of_points_y`, `nb_of_points_u`, `nb_of_points_v` depending on the
plot function being used. This might be confusing and it is definitely longer to type.

This commit add the capability to use `n1` and `n2` as the keyword arguments to specify
the number of discretization points in the first and second direction, respectively.

Back-compatibility is maintained.