Usability improvement
Added the capability to use `n` as the sole keyword argument to set the number of discretization points,
both for 2D and 3D plots. This is useful because many times the user doesn't really need to differentiate
between `n1` and `n2`: all the user wants his to increase the accuracy of the plot.

It is definitely quicker to type:

```
plot3d(cos(x**2 + y**2), n=1000)
```

than:

```
plot3d(cos(x**2 + y**2), n1=1000, n2=1000)
```