Beam: Increased compatibility for over-determined linear systems
This update causes inputs to the beam module to be automatically
converted to rationals when possible. This allows for many systems,
which were previously unsolvable due to rounding errors, to now
be solved exactly instead. However, this commit does not address the
general case of inconsistent systems (those without a solution).

For example, changing a beam object from `Beam(4, E, I)` to
`Beam(4.0, E, I)` no longer causes an error.