refactor(polys): move smith_normal_form to DomainMatrix
This commit moves the Smith Normal Code routine from Matrix to
DomainMatrix and replaces the original function with one that can
convert to DomainMatrix for the calculation.

The SNF code was the last place in the codebase that used RawMatrix so
the RawMatrix class is also removed. The RawMatrix class was a bad idea
because it subclasses Matrix in order to insert non-Expr elements
breaking a key internal invariant of the Matrix class. The removal of
RawMatrix is a step towards being able to enforce that the elements of a
Matrix should always be Expr.

This also adds tests to improve coverage in sympy.polys.matrices.