Integral(x, y).bound_variables -> [y]
The result of as_dummy for the evaluate-at
integral written as a limit of length 1 now
is written in the lenght-2 format to show
that something is both bound and free.

This singleton limit expression will now
be listed as a bound variable, too.

`doit` will not return an evaluated integral
when a function has symbols and functions with
any of those symbols in the bound symbols. e.g.
Integral(x + y, y) is not evaluated if y is f(x)
just like it is not evaluated when x = f(y).