integrals: fix definite integral of Heaviside function
Integrating the Heaviside function over an integral previously resulted
in some complicated meijerg functions, as mentioned in issue #21426.
Example:
In [1]: integrate(Heaviside(x-1), (x,-2,2))
Out[2]: -meijerg(((2, 1), ()), ((), (1, 0)), 1) + \
        meijerg(((2, 1), ()), ((), (1, 0)), -1) + \
        meijerg(((2, 1), ()), ((), (1, 0)), 2) - \
        meijerg(((2, 1), ()), ((), (1, 0)), -2)

Now, we instead get the simple and correct
Out[2]: 1

This also works if numbers are replaced by generic symbols, as done in
the test.

Note that the same could be achieved previously by passing
'meijerg=False' to the integration.

Also note that integration fails if anything but S.Half is passed on to
Heaviside as H0 argument, even though it should not affect the result of
the integration. This should need to be fixed in the future.

The code for the patch came from @jksuom.