fix(matrices): Don't use DomainMatrix.rref for RR in linsolve
The DomainMatrix.rref method as implemented by the sdm_irref function
has not been well designed for handling matrices of floats as it does
not have a pivoting strategy that can mitigate errors arising from
erroneously choosing a zero pivot (that is nonzero only because of
roundoff). This commit makes linsolve use Matrix.rref() for the case of
pure float matrices. The behaviour of sdm_rref for floats should be
improved. Until then it is better not to use DomainMatrix in these
cases.