perf(polys): make EX faster for trivial operations
This commit avoids calls to cancel for some trivial operations in the EX
domain. This can lead to a big speed up for some integrals. For example

z1, z2 = symbols('z1, z2')
I = Integral(exp(-0.5*(((-1/z1)+1.39)**2+((-1/z2)+1.23)**2))*(1/(z1**2))*(1/(z2**2)),(z1,0,1),(z2,0,1))
I.doit()

In this case the integral is not computed with or without this change
but heurisch completes much faster with the change reducing the overall
time by 50%. Previously most of the time was spent in trivial calls to
cancel like cancel(0), cancel(3), etc.