Make sinc(x).diff(x) not return a Piecewise
Currently, Piecewise.diff() is wrong for pointwise removable singularities.
This lead to sinc(x).diff(x, x) being wrong at x = 0. See
https://github.com/sympy/sympy/issues/11402.

The issue with Piecewise is difficult to solve, but we can work around sinc()
specifically by just returning cos(x)/x - sin(x)/x**2 for sinc(x).diff(x)
rather than a Piecewise. The value of the derivatives of sinc at 0 will have
to be evaluated with limit(), but this is preferable to having an incorrect
answer.