Pow._eval_as_leading term now raises when argument is unbounded Previously, Pow and exp _eval_as_leading_term methods returned the original expression itself when the argument was unbounded. Now, they raise a PoleError (example - exp(1/x).as_leading_term(x) raises an error and the code has accordingly been changed to account for this.
Due to the raising of an exception, the Order.__new__ code had to be
changed, as it relied on finding the leading term of the argument.
Now, it calculates the leading term of the args of its argument, and
depending on the type of argument, performs some simplifcation and
returns the answer.

fresnels and fresnelcs _eval_as_leading_term methods added