Fix the docs build when gmpy2 is installed
Previously the Sphinx would try to automatically generate a cross-reference to
gmpy2.mpz and gmpy2.mpq because of the dtype and tp attributes of the
GMPYIntegerRing and GMPYRationalField classes, which fails because we don't
have references to those classes in our docs. We don't really need to document
these, so we exclude them.

Ideally, we would be installing gmpy2 on CI when building the docs, which
would have caught this issue sooner, but this does not address that.

Fixes #21697.