Don't try to guard against non-hashable results in the cache
See https://github.com/sympy/sympy/pull/21619#issuecomment-889540868. This can
cause issues when the function itself raises TypeError. Really, if a function
has the cacheit decorator and returns a non-hashable result, that is a bug.
Either the function should be fixed to return a hashable result, or some lower
level function that does always return hashable results should be cached
instead.

This commit is to see what sorts of tests will fail when this change is made.
We may also want to run the test suite in random order to try to find
failures, since that can affect caching.