Guard against the *input* being non-hashable in the cache
The previous code assumed any TypeError was a hashing problem, meaning it
would run the code twice when that happened. The code in the caching is based
around the output needing to be hashable, but this is incorrect (I haven't
fixed that yet). This commit is primarly to see if this leads to any test
failures.