Removed unnecessary normalization in quaternion.py
removed unnecessary normalization of the created quaternion in the from_axis_angle Quaternion method.  The quaternion is normalized by construction.

Also added an is_normalized flag to Quaternion.rotate_point, which allows the user to skip the forced normalization of the input quaternion, if the user knows that it is already normalized.

Forgoing these normalizations is sometimes a key operation to be able to successfully simplify complex algebraic output from these routines.