keywords give input; no symbols
It seems the most natural to give input regarding the
set elements as opposed to the multiplicity of element
repetition (though I suppose it depends on who is using
the function). So specifying n distinct items or multiplicities
of element counts require use of keywords n and m to pass
that information, respectively.

nD('123') == nD(n=3) == nD(m=(1,1,1)) == nD(m={1:3}) == 2