Fix help(latex) to show the docstring
Previously, functools.update_wrapper() was used on the wrapper class instance.
But this added __doc__ to the class instance, which does not work with help()
(help() only looks at __doc__ if it is defined on the class). So we instead
redefine the wrapper class for each decorated function, and use update_wrapper
on the class rather than the instance).

Use a dynamically created type for _PrintFunction in Python <3.8

This is simpler than moving the class definition to the body of the function.
__doc__ does not need to be on the class itself in Python 3.9 for it to work
with help(), but it does for 3.8 and lower. This alternative change was
suggested by @eric-wieser.

This also removes __repr__ from the class, as it only serves to make things
confusing when debugging this class, and there's no reason to completely hide
the fact that it is a class instead of the function.

Co-authored-by: Aaron Meurer <asmeurer@gmail.com>
Co-authored-by: Oscar Gustafsson <oscar.gustafsson@gmail.com>