fix(core): improve checks in pow substitutions
In a substitution like (1/x).subs(x**2, y) there need to be checks to
see whether it is valid to replace x with e.g. sqrt(y). Previously those
checks were heuristic and depended on the evaluation behaviour of
_eval_power. This commit changes those to depend directly on the
assumptions of the base and exponent in the base pow. This change fixes
a number of incorrect results from manual integration.