closes #22077
signsimp during match changes structure of expressions
which breaks match expectations. The tail wagged the dog,
I think, on the addition of signsimp to solve a problem
with `solve`'s use of match. Match is no longer used in
`solve` and test expressions for matching are working as
exected in tests that were added in #22077.