fix(polys): convert mpz to int when gmpy is not used
This commit ensures that e.g. ZZ.sqrt(2) will be of type int when gmpy
is not used. The problem is that although SymPy might not use gmpy in
all circumstances mpmath might and so SymPy needs to convert the output
of mpmath functions if it is not also using gmpy.