fix(printing): suppress printing default Heaviside arg
Heaviside(x) was changed to have a default value of 1/2 for the second
argument rather than leaving it undefined. This changed the args from
(x,) to (x, 1/2) which then changed the output of several printers. This
commit restores the previous printing behaviour so that if the second
arg for Heaviside(x, y) is equal to the default 1/2 then it is
suppressed from printing in the str/repr/latex/octave printers.