maint(release): don't guess previous version
The release script attempts to identify new authors by guessing the
correct tag for the previous SymPy release but this fails under a number
of scenarios such as if the current release tag is for a merge commit or
if the previous release tag has not ultimately been merged into master.
This commit prevents trying to guess the tag and instead requires the
previous version to be provided explicitly.