less trivial changes
The second arg of laplace_transform should tell the range
in which the transform is valid and the 3rd arg gives
additional conditions. I *think* these are ok if the range
re(s) > 0 (a=0 as second arg) replaces a=x for the second
arg and a 3rd arg consistent with s > x...But not everything
is that simple in terms of changes. These changes come about
from adding rules in _condsimp. So if anything there is wrong
then it will cause errors in the final answer.