CodeParsers to preprocess expressions for CodePrinters
SymPy features a very extensive printing system. In this sytem, the `doprint`
method of an appropriate `Printer` is used to obtain a string. Internally,
doprint will use the `_print` method which looks for an appropriate
`_print_<class>` based on the class (or its parent classes) of the object
being printed. Additionally, an object can define its own printing method
by giving it the appropriate name (Printer dependent) such that users
can define the printing behaviour for their custom objects. These
`_print_<class>` methods will often call `_print` again. For example: `Mul`,
the object in SymPy representing multiplication, might use `_print` on all
of its arguments and then join the obtained strings with a "*" character.

Here, the `Parser` class is introduced which generalises the above concepts.
The `Parser` class allows the use of not only `doprint`, `_print` and
`_print_<class>`, but now allows print to be replaced with any name (without
underscores "_") using `do<parsername>`, `_<parsername>` and
`_<parsername>_<class>`. Multiple such parsers can be stored in the same
object since they should not conflict. Additionally, `do<parsername>` and
`_<parsername>`	are automatically constructed if they are not provided.

Currently, several parts of SymPy rely heavily on the code printers. For
example, `lambdify` and `autowrap` use the code printers to convert SymPy
expressions into code that can be run in several different languages.
However, both use their own code generation to obtain the code needed to
embed the printed expression in a form that is expected of a programming
language (e.g. function definitions, variable declarations etc.). Code
parsers could be used to perform these tasks, such that new functionality
is immediately available in both `autowrap` and `lambdify` as well as to
the user simply wanting to obtain the code and reuse it in their own
projects.

Recently, common subexpression extraction (CSE) was implemented in
`lambdify`, which was also used to delete temporary variables to reduce
the memory usage. Automatic (de)allocation would also be useful for
`autowrap` but would now need to be implemented again. On top of that,
putting new functionality in code parsers would stop bloating the already
lengthy lambdify and autowrap functions.

Examples of how to use code parsers to follow.