fixes #21890
The equation that is solved ends up sending an expression which
simplifies to basically I/x with x real so it fails to compare.
So the _is_connected routine now returns False immediately when
that is detected.

Curiously, Relational(floor(x), x, '<=') does not evaluate,
nor does is_le(floor(x), x), even when x is real, but the
expression floor(x) <= x does evaluate, so that sort of literal
expression is retained in the _is_connected routine.

>>> Relational(floor(x),x,'<=')
floor(x) <= x
>>> is_le(floor(x),x)
>>> floor(x) <= x
True
>>> x >= floor(x)
x >= floor(x)