Reorganize `polys.numberfields`.
The set of problems solved in `polys.numberfields` can be organized in a better way, which the present module structure fails to reflect.
Note also: the present structure was only put in effect into a recent refactoring (since the last release).

What we currently have in `polys.numberfields` is:
1. A way to compute minimal polynomials for algebraic numbers.
2. A solution to the Subfield Problem and its several allied problems (see [Cohen93], Section 4.5, or the docstring for the new `subfield.py` module).
3. An ad hoc function called `isolate()`, for returning an isolating interval for a real algebraic number.

Accordingly, we should have three modules:
1. minpoly.py
2. subfield.py
3. utilities.py

Therefore this commit makes the following moves:

* Rename `isomorphism.py` --> `subfield.py`
* Move `primitive_element()` function from `minpoly.py` --> `subfield.py`
* Move `to_number_field()` function from `numbers.py` --> `subfield.py`
* Rename `numbers.py` --> `utilities.py`
* Move unit tests in a corresponding way.