Convert svg files to pdf in the latex docs build using headless Chrome
Previously this used inkscape, but it was misconverting several of the SVG
images, in some cases in ways that made the image completely broken. There are
several other possibilities (see
https://superuser.com/questions/381125/how-do-i-convert-an-svg-to-a-pdf-on-linux#),
including rsvg-convert, convert, and cairosvg, but all have similar issues.
Additionally, inkscape is a pain to install (at least on Mac), and newer
versions of inkscape changed the command line flags, breaking the Makefile.

A script convert-svg-to-pdf.sh is added that uses headless Chrome or Chromium
to convert the SVG files to PDF, which is based on this gist
https://gist.github.com/s417-lama/84bf66de1096c4587e8187092fb41684.

Note that this is only require for building the PDF docs (which also requires
LaTeX). The HTML documentation uses the SVG files directly and does not
require any conversions. Inkscape is no longer a build dependency for the
documentation.