Fix benchmarks build on GitHub Actions
The build is broken because of an argparse bug introduced in Python 3.9.8,
which breaks asv machine --config (it ignores the given filename and only
looks at the default file). This won't be fixed until 3.9.10, so for now, we
work around this by renaming asv.conf.actions.json to the expected
asv.conf.json. Once 3.9.10 is released and Actions is updated, we can revert
this commit. Travis also runs the benchmarks, but it is still using 3.9.1, so
this issue is not present there (if Travis started failing that wouldn't be a
huge deal anyway).

Fixes #22459.