Bugfix in `_minpoly_groebner()`.
Expressions raised to positive integer powers were incorrectly handled.

New unit test shows a case that was failing before.

This case was being handled the same way as root extractions. This meant that, e.g.,
if we had `a0**3`, with `a0` an algebraic number, then we would introduce a new symbol
`a1` with the defining "polynomial" `a1**(1/3) - a0`. Since this is not actually a polynomial,
we crashed when trying to compute the Groebner basis.

We also add some doctext to explain how `_minpoly_groebner()` works.