Bugfix in `minimal_polynomial()` and `field_isomorphism()`.
The `minimal_polynomial()` and `field_isomorphism()` functions were not always doing what they advertised to do.
This was due to a misapprehension of what an instance of the `AlgebraicNumber` class represents.

An `AlgebraicNumber` instance represents an element `alpha` of a number field `Q(theta)`. The primitive
element `theta` for the number field is represented by its minimal polynomial, and a particular root thereof.
The element `alpha` is represented as a polynomial in `theta`.

The `minimal_polynomial()` and `field_isomorphism()` functions were written as if an `AlgebraicNumber`
instance represented not `alpha` but `theta`. In other words, these functions operated on the primitive
element of the number field, not on the represented element itself.

We repair this as follows:
* Provide new docstrings for the `AlgebraicNumber` class, explaining what it represents and how.
* Provide new methods for the `AlgebraicNumber` class, making it easy to:
  - get the minimal polynomial for alpha
  - test whether alpha == theta
  - convert into an instance were alpha does equal theta
* Update the `minimal_polynomial()` and `field_isomorphism()` functions to operate on alpha, not theta.
* Provide new unit tests that failed before the fix.