Add autosummary class template
We unfortunately cannot use the default template, which recursively uses
autosummary on all the methods of the class, because this just documents every
method of the class, including methods inherited from Basic, private "dunder"
methods, and methods with no docstrings, and there is no way to configure
this. It can't even be configured in the template because the autosummary
template variables don't know about the autodoc distinctions for things like
inherited-members.

One option in the future may be to use autodocsumm instead of autosummary,
which does know about the various autodoc distinctions. Unfortunately,
autodocsumm does not support putting the documentation on separate pages,
which is the whole point of our using autosummary in the first place. See
https://github.com/Chilipp/autodocsumm/issues/66.