Revise `AlgebraicNumber.__new__()`.
Do not make the conversion `r = r.to_root()` in the case that the
first arg is a pair `(m, r)`, with `r` an `AlgebraicNumber` instance.

This was an attempt to address an issue with pickling of AlgebraicNumber and
AlgebraicField intances, but it causes problems with the preservation of
aliases. We will open this in a separate issue, and handle in a separate PR.