sort expressions when replacing with Dummy
If signsimp is ever used during processing of an expression in which
Dummy has been used to replace expressions, the expressions must be
assigned Dummy symbols in canonical order or else expressions that were
in one order might differ from the order they would be in when replaced
with Dummy. e.g.

    >>> eq = 1/(cos(x) - cos(y))
    >>> assert signsimp(eq) == eq
    >>> d1,d2 = Dummy(), Dummy()
    >>> assert signsimp(1/(d1-d2)) == -signsimp(1/(d2-d1))

So if eq is to remain unchanged after signsimp (and d1 and d2 are going
to replace the cos terms) then the order of assigment has to match the
ordering of the cos terms.